some more adjustements
