+++ LuckParser/Models/ParseModels/Players/AbstractPlayer.cs
@@ -41,16 +41,6 @@ namespace LuckParser.Models.ParseModels
         {
             return agent.getProf();
         }
-        public long getDeath(ParsedLog log, long start, long end)
-        {
-            long offset = log.getBossData().getFirstAware();
-            CombatItem dead = log.getCombatList().LastOrDefault(x => x.getSrcInstid() == agent.getInstid() && x.isStateChange().IsDead() && x.getTime() >= start + offset && x.getTime() <= end + offset);
-            if (dead != null && dead.getTime() > 0)
-            {
-                return dead.getTime();
-            }
-            return 0;
-        }
 
         public List<DamageLog> getDamageLogs(int instidFilter, ParsedLog log, long start, long end)//isntid = 0 gets all logs if specefied sets and returns filterd logs
         {
